Initial commit
Project for Bundling Azure Files. Minify and compress groups of files
from Azure blob storage. Create "Bundles" to be served from Azure CDN
with a fallback to cached in-memory Bundle. Uses a Cache Dependency to
invalidate Bundles. Works just like regular ASP.NET bundling and
minification but uses files from Azure instead of the file system.
Creates a minified file and a minified/compressed file in Azure blob
storage to be served to the client.